+++ CargoHubV2/app/controllers/items_controller.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter, HTTPException, Depends
 from sqlalchemy.orm import Session
-from CargoHubV2.app.schemas.items_schema import *
 from CargoHubV2.app.database import get_db
-from CargoHubV2.app.services.items_service import *
+from CargoHubV2.app.schemas.item_schema import *
+from CargoHubV2.app.services.item_service import *
 from typing import Optional, List
